+++ RELEASE-NOTES.md
@@ -1,9 +1,12 @@
 ## RELEASE NOTES
+### Version 5.0.22-case-file-view-update-document-category
+**EUI-6815** Service layer for HTTP PUT call to CCD endpoint for updating the category associated with a document
 
 ### Version 5.0.22-case-file-view-two-sections
 **EUI-6822** Two sections for case file view
 **EUI-5376** Slider function for case file view
 
+### Version 5.0.22-case-file-view-get-categories-and-documents
 **EUI-5481** Service layer for HTTP GET call to CCD endpoint for retrieving categories and documents
 
 ### Version 5.0.22-angular11-upgrade